added retrofit client

+++ app/src/main/java/com/kirillmesh/vknewsclient/domain/StatisticElement.kt
@@ -6,7 +6,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class StatisticElement (
     val type: StatisticType,
-    val count: Int = 0
+    val count: Long = 0
 ) : Parcelable
 
 enum class StatisticType{
